refactor(visorconfig,dmsg/disc): drop net/http from WASM build graph
Before: visorconfig and pkg/dmsg/disc each imported net/http for
runtime-only code paths (visorconfig.Fetch + CookieConfig.SameSite;
disc.NewHTTP + HTTPMessage status-code constants + mock client).
That pulled net/http into the WASM build graph used by
pkg/skywireconfig/{autoconfigcmd,keypair,genvisor} → apt-repo's
install-page WASM.

The net/http inclusion was harmless for the Go toolchain but
TinyGo 0.41.1's stdlib has a roundTrip case-mismatch in
net/http/roundtrip_js.go that surfaces once the transitive
import surface widens (e.g. once genvisor pulls in
deployment.Prod's services-config JSON unmarshal path). TinyGo
builds were failing on every refresh.

This refactor splits each affected package into a "WASM-clean"
core file and a //go:build !js extension file:

  pkg/dmsg/disc/
    interface.go         (new, no build tag) — EntryReader,
                         EntryWriter, APIClient interfaces +
                         the package-wide jsoniter `json` var.
    entry.go             (no build tag, no net/http) — Entry
                         struct, Marshal/Unmarshal logic.
    fallback.go          (no build tag) — wraps two APIClients.
    client.go            (//go:build !js) — httpClient impl.
    http_message.go      (//go:build !js) — HTTPMessage status
                         consts.
    testing.go           (//go:build !js) — mock APIClient.

  pkg/visor/visorconfig/
    services.go          (no build tag, no net/http) — Services
                         + EnvServices types + HasDmsgEndpoints.
    services_native.go   (new, //go:build !js) — Fetch helper.
    hypervisorconfig.go  (no build tag, no net/http) — config
                         types minus SameSite.
    hypervisorconfig_native.go (new, //go:build !js) —
                         CookieConfig.SameSite.

After: `GOOS=js GOARCH=wasm go list -deps ./pkg/skywireconfig/...
./pkg/visor/visorconfig/` no longer mentions net/http or any of
its 7 transitive net/http/* internal packages. Go WASM build
size drops from 8.6MB → 7.3MB (≈ −15%).

TinyGo still fails for a separate reason — reflect runtime
helpers (reflect.unsafe_New, reflect.mapassign, ...) needed by
encoding/json. Removing encoding/json from the WASM graph is a
larger refactor (logrus, google/uuid, blang/semver, jsoniter all
import it transitively). Out of scope for this PR; the Go-WASM
path remains the production install-page artifact.

Verified:
  go build ./...                                    — clean
  GOOS=js GOARCH=wasm go build ./pkg/{skywireconfig/...,
    visor/visorconfig/,dmsg/disc/}                  — clean
  go test ./pkg/{dmsg/disc/,visor/visorconfig/,
    skywireconfig/...}                              — pass
