+++ pkg/dmsg/disc/client.go
@@ -1,4 +1,12 @@
+//go:build !js
+
 // Package disc pkg/disc/client.go
+//
+// HTTP-based APIClient. Build-tag-gated to keep net/http out of the
+// js/wasm build graph; the EntryReader/EntryWriter/APIClient
+// interfaces live in interface.go (no build tag) so consumers like
+// pkg/visor/visorconfig still see the type names under GOOS=js
+// without dragging net/http in. See interface.go for the rationale.
 package disc
 
 import (
@@ -10,36 +18,9 @@ import (
 	"sync"
 	"time"
 
-	jsoniter "github.com/json-iterator/go"
-
 	"github.com/skycoin/skywire/pkg/cipher"
 	"github.com/skycoin/skywire/pkg/logging"
 )
-
-var json = jsoniter.ConfigFastest
-
-// EntryReader provides read-only access to discovery entries.
-type EntryReader interface {
-	Entry(context.Context, cipher.PubKey) (*Entry, error)
-	AvailableServers(context.Context) ([]*Entry, error)
-	AllServers(context.Context) ([]*Entry, error)
-}
-
-// EntryWriter provides write access to discovery entries.
-type EntryWriter interface {
-	PostEntry(context.Context, *Entry) error
-	PutEntry(context.Context, cipher.SecKey, *Entry) error
-	DelEntry(context.Context, *Entry) error
-}
-
-// APIClient implements dmsg discovery API client.
-type APIClient interface {
-	EntryReader
-	EntryWriter
-	AllEntries(ctx context.Context) ([]string, error)
-	AllClientsByServer(ctx context.Context) (map[string][]*Entry, error)
-	ClientsByServer(ctx context.Context, serverPK cipher.PubKey) ([]*Entry, error)
-}
 
 // HTTPClient represents a client that communicates with a dmsg-discovery service through http, it
 // implements APIClient
